sbalance: Allow IRQs to be moved off of excluded CPUs
Excluded CPUs are excluded from IRQ balancing with the intention that those
CPUs shouldn't really be processing interrupts, and thus shouldn't have IRQs
moved to them. However, SBalance completely ignores excluded CPUs, which can
cause them to end up with a disproportionate amount of interrupt traffic that
SBalance won't spread out. An easy example of this is when CPU0 is an excluded CPU,
since CPU0 ends up with all interrupts affined to it by default on arm64.

Allow SBalance to move IRQs off of excluded CPUs so that they cannot slip
under the radar and pile up on an excluded CPU, like when CPU0 is excluded.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>